Revert "Secure API dev"
This reverts commit 320fef9f4bb325df23c5159247d794fdaed9b183.
